Preventing Access: The Easy Way
  Preventing me from eating a cookie baked by someone else is actually pretty simple. And what we should do first is just put the logic into our controller. So I’ll do that here: if the baker’s username does not equal the current user’s username, we’re going to throw that AccessDeniedException and say: 'Hey you didn’t bake this! '

  Now if we try to eat one of Leanna’s cookies she catches us with a nice clear message. And of course in the production environment, this would be your 403 error page.

Tip
See Error Pages(https://knpuniversity.com/screencast/symfony2-ep3/error-pages) for how to customize your 404, 403 and 500 error pages in production.

+++ src/AppBundle/Controller/CookieController.php
@@ -27,11 +27,16 @@ class CookieController extends Controller {
    */
   public function nomAction(Request $request, $id) {
     $em = $this->getDoctrine()->getManager();
+    
     $cookie = $em->getRepository('AppBundle:DeliciousCookie')
       ->find($id);
-
+    
     if (!$cookie) {
       throw $this->createNotFoundException();
+    }
+    
+    if($cookie->getBakerUsername() != $this->getUser()->getUsername()){
+      throw $this->createAccessDeniedException('You didn\'t baked this!');
     }
 
     $em->remove($cookie);
